resolved pr changes and completed download test draft integration
