--- src/TestCases/FunctionMaster/Validation/AddFunctionMaster.js
@@ -1,6 +1,0 @@
-import * as Yup from "yup";
-
-export const addFunctionMasterValidation = Yup.object().shape({
-  title: Yup.string().required("Function title is required"),
-  remark: Yup.string(),
-});
