--- src/TestCases/TestDraft/Validation/DownloadFormatFile.js
@@ -1,7 +1,0 @@
-import * as Yup from "yup";
-
-export const downloadFormatFile = Yup.object().shape({
-  project_name: Yup.string().required("Function is required"),
-  module_name: Yup.string().required("Function is required"),
-  submodule_name: Yup.string().required("Function is required"),
-});
