+++ src/screens/TestCases/TestDraft/Validation/DownloadFormatFile.js
@@ -1,0 +1,6 @@
+import * as Yup from "yup";
+
+export const downloadFormatFile = Yup.object().shape({
+  project_id: Yup.string().required("Project Name is required"),
+  module_id: Yup.string().required("Module Name is required"),
+});
